+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Next
 * Added Travis-CI - [@dblock](https://github.com/dblock).
 * Added TODO list and enhanced README documentation - [@ptitvert](https://github.com/ptitvert)
 * Your contribution here.
-* Added support for CPU load in OSX [@kamenitxan](https://github.com/kamenitxan)
+* Added support for CPU load in OSX and Linux[@kamenitxan](https://github.com/kamenitxan)
 
 
 1.1 (10/13/2013)